Add AppStream release data
Release information is a part of the stricter AppStream validation
requirements.

Although it is not a submission requirement, Flathub aims to include
this metadata for all applications. With this data Flathub and software
centers can display the version and date of apps. Additionally,
simple to understand, changelogs can be presented prior to updates.