Add Travis CI build tests
These files can be used to setup Travis CI, in a way that tests every
commit to the pulseeffects repo.

Tests are run using the same PKGBUILD that builds "pulseeffects-git".
That way things like dependencies only have to be maintained in a single
file, opposed to using the functionality offered by .travis.yml
(and arch-travis). It should also the most comfortable and familiar
option to the lead developer.

Travis CI has no native support for Arch Linux, but it does have an
option to run a virtual machine, with docker containers. To compensate
for this lack of support, a modified version of mikkeloscar/arch-travis
is used.