Use existing $srcdir with makepkg
When `makepkg` is executed it uses current directory as its working
directory. Because of this it will store files in ./src, ./pkg and
./pulseeffects.

This is not only annoying to cleanup, it has a chance of conflicting
with existing files and directories. The solution is to move the
PKGBUILD to a dedicated sub directory.

In addition this commit will use the existing source directory, instead
of downloading a second copy of:

- https://github.com/wwmm/pulseeffects.git

Just be careful not to use any operations that delete or cleanup $srcdir
with makepkg, if testing from your work repo.