Add manual user documentation kudo
The AppStream generator that processes the appdata for use with
front-ends like Gnome Software fails to detect any user documentation.

Since PulseEffects does in fact provide very good documentation it is
acceptable to set this manually. See:
https://gitlab.gnome.org/GNOME/gnome-software/blob/master/doc/kudos.md#manually-adding-kudos

This fix ensures that Gnome Software (and others?) properly informs
users that this app includes good documentation, which might influence
the decision whether to install the app.